requirements.txt
djangorestframework
djangorestframework-simplejwt
djangorestframework-jwt
psycopg2-binary
gunicorn
dj_rest_auth
django-cors-headers
bcrypt
PILLOW